element

+++ NonPerishable.java
@@ -1,18 +1,18 @@
 public class NonPerishable extends Product{
     
-    private int volume;
+    private double volume;
 
-    public NonPerishable(String name, double cost,int volume) {
+    public NonPerishable(String name, double cost,double volume) {
         
         super(name, cost);
         setVolume(volume);
     }
 
-    public int getVolume() {
+    public double getVolume() {
         return volume;
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
